api moi nhat 2

+++ Controllers/ProductController.cs
@@ -1,7 +1,9 @@
 ﻿using BackEnd.Attributes;
+using BackEnd.Extensions;
 using BackEnd.Filters;
 using BackEnd.Models;
 using BackEnd.Services;
+using BackEnd.ViewModels;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,25 +14,56 @@ namespace BackEnd.Controllers
     public class ProductController : ControllerBase
     {
         private readonly IProductService _productService;
-
-        public ProductController(IProductService productService)
+        private readonly IProductReturnService _productReturnService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public ProductController(IProductService productService, IHttpContextAccessor httpContextAccessor, IProductReturnService productReturnService)
         {
             _productService = productService;
+            _httpContextAccessor = httpContextAccessor;
+            _productReturnService = productReturnService;
         }
+
+        //https://localhost:7002/api/product/AddReturnProduct
+        //
+        [HttpPost("AddReturnProduct")]
+        [StaffAuthorize]
+
+        public async Task<IActionResult> AddReturnProduct(ProductReturnViewModel product)
+        {
+            int? staff = _httpContextAccessor.HttpContext.GetStaffId();
+            if (staff == null)
+            {
+                return Unauthorized("Session has expired");
+            }
+            var result = await _productReturnService.CreateProductReturnAsync(product);
+            if (result.IsNewProduct)
+            {
+                return Ok(new { Message = "New product added", Product = result.Product });
+            }
+            else
+            {
+                return Ok(new { Message = "Product return processed", ProductReturn = result.ProductReturn });
+            }
+        }
+
+
         //https://localhost:7002/api/product/add
         //ok
         [HttpPost("add")]
         [ProductValidationFilter]
-        [AdminAuthorize]
-        
+        //      [AdminAuthorize]
+
         public async Task<IActionResult> AddProduct( Product product)
         {
+            int? admin = _httpContextAccessor.HttpContext.GetAdminId();
+            if (admin == null) { throw new Exception("phien dang nhap het han"); }
             var newProduct = await _productService.AddProductAsync(product);
             return Ok(newProduct);
         }
-
+        //htt
         //https://localhost:7002/api/product/search/{id}
         //ok
+        //get lay tu trong du lieu database
         [HttpGet("search/{id}")]
         public async Task<IActionResult> GetProduct(int id)
         {
@@ -52,9 +85,12 @@ namespace BackEnd.Controllers
         //https://localhost:7002/api/product/update/{id}
         //ok
         [HttpPut("update/{id}")]
-        [AdminAuthorize]
+        //      [AdminAuthorize]  
+        //put update trong database
         public async Task<IActionResult> UpdateProduct(int id, [FromBody] Product product)
         {
+            int? admin = _httpContextAccessor.HttpContext.GetAdminId();
+            if (admin == null) { throw new Exception("phien dang nhap het han"); }
             var updated = await _productService.UpdateProductAsync(id, product);
             if (!updated)
             {
@@ -65,9 +101,11 @@ namespace BackEnd.Controllers
         //https://localhost:7002/api/product/delete/{id}
         //ok
         [HttpDelete("delete/{id}")]
-        [AdminAuthorize]
+        //      [AdminAuthorize]
         public async Task<IActionResult> DeleteProduct(int id)
         {
+            int? admin = _httpContextAccessor.HttpContext.GetAdminId();
+            if (admin == null) { throw new Exception("phien dang nhap het han");}
             var deleted = await _productService.DeleteProductAsync(id);
             if (!deleted)
             {
